requirements.txt
Flask==1.1.2
WTForms==2.3.3

Pillow~=9.4.0